Move interactive example server related classes into examples.server package in core Support generation of multiple examples for a single endpoint
